improve test
