+++ datafusion/core/tests/sql/parquet_schema.rs
@@ -100,6 +100,84 @@ async fn schema_merge_ignores_metadata_by_default() {
 }
 
 #[tokio::test]
+async fn merge_compatible_schemas() -> Result<()> {
+    let tmp_dir = TempDir::new()?;
+
+    let schema1 = Schema::new_with_metadata(
+        vec![
+            Field::new("a", DataType::Int8, false),
+            Field::new("b", DataType::Int16, false),
+            Field::new("c", DataType::Int32, false),
+        ],
+        HashMap::new(),
+    );
+
+    let schema2 = Schema::new_with_metadata(
+        vec![
+            Field::new("a", DataType::Int64, false),
+            Field::new("b", DataType::Int32, false),
+            Field::new("c", DataType::Int16, false),
+        ],
+        HashMap::new(),
+    );
+
+    create_parquet_file(tmp_dir.path(), "t1.parquet", &schema1).await?;
+    create_parquet_file(tmp_dir.path(), "t2.parquet", &schema2).await?;
+
+    let ctx = SessionContext::default();
+    ctx.register_parquet(
+        "test",
+        tmp_dir.path().to_str().unwrap(),
+        ParquetReadOptions::default(),
+    )
+    .await?;
+    let df = ctx.table("test")?;
+
+    let expected_schema = Schema::new_with_metadata(
+        vec![
+            Field::new("a", DataType::Int64, false),
+            Field::new("b", DataType::Int32, false),
+            Field::new("c", DataType::Int32, false),
+        ],
+        HashMap::new(),
+    );
+
+    let actual_schema: Schema = df.schema().into();
+    assert_eq!(expected_schema, actual_schema);
+
+    let batches = df.collect().await?;
+    for batch in batches {
+        assert_eq!(&expected_schema, batch.schema().as_ref());
+        assert!(batch
+            .column(0)
+            .as_any()
+            .downcast_ref::<Int64Array>()
+            .is_some());
+        assert!(batch
+            .column(1)
+            .as_any()
+            .downcast_ref::<Int32Array>()
+            .is_some());
+        assert!(batch
+            .column(2)
+            .as_any()
+            .downcast_ref::<Int32Array>()
+            .is_some());
+    }
+
+    Ok(())
+}
+
+async fn create_parquet_file(path: &Path, filename: &str, schema: &Schema) -> Result<()> {
+    let ctx = SessionContext::default();
+    let options = CsvReadOptions::new().schema(&schema);
+    ctx.register_csv("t", "tests/example.csv", options).await?;
+    let t = ctx.table("t")?;
+    let path = path.join(filename);
+    t.write_parquet(path.to_str().unwrap(), None).await
+}
+
+#[tokio::test]
 async fn schema_merge_can_preserve_metadata() {
     // Create several parquet files in same directoty / table with
     // same schema but different metadata
